LPS-124006 Embed ImagePropertiesPanel inside ImageSourcePanel
