+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/item-configuration-panels/ImageSourcePanel.js
@@ -15,11 +15,24 @@
 import ClayForm, {ClaySelectWithOption} from '@clayui/form';
 import React, {useState} from 'react';
 
-import {useSelector} from '../../../../app/store/index';
+import {BACKGROUND_IMAGE_FRAGMENT_ENTRY_PROCESSOR} from '../../../../app/config/constants/backgroundImageFragmentEntryProcessor';
+import {EDITABLE_FRAGMENT_ENTRY_PROCESSOR} from '../../../../app/config/constants/editableFragmentEntryProcessor';
+import {EDITABLE_TYPES} from '../../../../app/config/constants/editableTypes';
+import {VIEWPORT_SIZES} from '../../../../app/config/constants/viewportSizes';
+import {config} from '../../../../app/config/index';
+import selectEditableValueContent from '../../../../app/selectors/selectEditableValueContent';
+import selectLanguageId from '../../../../app/selectors/selectLanguageId';
+import selectSegmentsExperienceId from '../../../../app/selectors/selectSegmentsExperienceId';
+import {useDispatch, useSelector} from '../../../../app/store/index';
+import updateEditableValuesThunk from '../../../../app/thunks/updateEditableValues';
 import isMapped from '../../../../app/utils/isMapped';
+import {setIn} from '../../../../app/utils/setIn';
+import {updateIn} from '../../../../app/utils/updateIn';
 import {useId} from '../../../../app/utils/useId';
-import getEditableItemPropTypes from '../../../../prop-types/getEditableItemPropTypes';
-import {ImagePropertiesPanel} from './ImagePropertiesPanel';
+import {ImageSelector} from '../../../../common/components/ImageSelector';
+import {ImageSelectorDescription} from '../../../../common/components/ImageSelectorDescription';
+import {ImageSelectorSize} from '../../../../common/components/ImageSelectorSize';
+import {getEditableItemPropTypes} from '../../../../prop-types/index';
 import {MappingPanel} from './MappingPanel';
 
 const SOURCE_OPTIONS = {
@@ -37,6 +50,10 @@ export default function ImageSourcePanel({item}) {
 	const fragmentEntryLinks = useSelector((state) => state.fragmentEntryLinks);
 	const sourceSelectionInputId = useId();
 
+	const selectedViewportSize = useSelector(
+		(state) => state.selectedViewportSize
+	);
+
 	const editableValue =
 		fragmentEntryLinks[item.fragmentEntryLinkId].editableValues[
 			item.editableValueNamespace
@@ -48,34 +65,218 @@ export default function ImageSourcePanel({item}) {
 			: SOURCE_OPTIONS.direct.value
 	);
 
-	const ConfigurationPanel =
-		source === SOURCE_OPTIONS.direct.value
-			? ImagePropertiesPanel
-			: MappingPanel;
+	let ConfigurationPanel = DirectImagePanel;
+
+	if (source === SOURCE_OPTIONS.mapping.value) {
+		if (selectedViewportSize === VIEWPORT_SIZES.desktop) {
+			ConfigurationPanel = MappingPanel;
+		}
+		else {
+			ConfigurationPanel = null;
+		}
+	}
 
 	return (
 		<>
-			<ClayForm>
-				<ClayForm.Group>
-					<label htmlFor={sourceSelectionInputId}>
-						{Liferay.Language.get('source-selection')}
-					</label>
+			{selectedViewportSize === VIEWPORT_SIZES.desktop && (
+				<ClayForm>
+					<ClayForm.Group>
+						<label htmlFor={sourceSelectionInputId}>
+							{Liferay.Language.get('source-selection')}
+						</label>
 
-					<ClaySelectWithOption
-						className="form-control form-control-sm mb-3"
-						id={sourceSelectionInputId}
-						onChange={(event) => setSource(event.target.value)}
-						options={Object.values(SOURCE_OPTIONS)}
-						value={source}
-					/>
-				</ClayForm.Group>
-			</ClayForm>
+						<ClaySelectWithOption
+							className="form-control form-control-sm mb-3"
+							id={sourceSelectionInputId}
+							onChange={(event) => setSource(event.target.value)}
+							options={Object.values(SOURCE_OPTIONS)}
+							value={source}
+						/>
+					</ClayForm.Group>
+				</ClayForm>
+			)}
 
-			<ConfigurationPanel item={item} />
+			{ConfigurationPanel && <ConfigurationPanel item={item} />}
 		</>
 	);
 }
 
 ImageSourcePanel.propTypes = {
+	item: getEditableItemPropTypes().isRequired,
+};
+
+function DirectImagePanel({item}) {
+	const {editableId, fragmentEntryLinkId, type} = item;
+
+	const dispatch = useDispatch();
+	const editables = useSelector((state) => state.editables);
+	const fragmentEntryLinks = useSelector((state) => state.fragmentEntryLinks);
+	const languageId = useSelector(selectLanguageId);
+	const segmentsExperienceId = useSelector(selectSegmentsExperienceId);
+	const selectedViewportSize = useSelector(
+		(state) => state.selectedViewportSize
+	);
+
+	const processorKey =
+		type === EDITABLE_TYPES.backgroundImage
+			? BACKGROUND_IMAGE_FRAGMENT_ENTRY_PROCESSOR
+			: EDITABLE_FRAGMENT_ENTRY_PROCESSOR;
+
+	const editableValues =
+		fragmentEntryLinks[fragmentEntryLinkId].editableValues;
+
+	const editableValue = editableValues[processorKey][editableId];
+	const editableConfig = editableValue.config || {};
+	const editableElement = editables?.[item.parentId]?.[item.itemId]?.element;
+
+	const editableContent = selectEditableValueContent(
+		{fragmentEntryLinks, languageId},
+		fragmentEntryLinkId,
+		editableId,
+		processorKey
+	);
+
+	const imageUrl =
+		typeof editableContent === 'string'
+			? editableContent
+			: editableContent?.url;
+
+	const imageTitle =
+		editableConfig.imageTitle ||
+		(imageUrl === editableValue.defaultValue ? '' : imageUrl);
+
+	const imageDescription =
+		typeof editableConfig.alt === 'object'
+			? editableConfig.alt[languageId] ||
+			  editableConfig.alt[config.defaultLanguageId] ||
+			  ''
+			: editableConfig.alt || '';
+
+	const imageSizeId = editableConfig.imageProperties?.[selectedViewportSize];
+
+	const handleImageChanged = (nextImage) => {
+		const nextEditableValue = {
+			...editableValue,
+
+			config: {
+				alt: {},
+				imageConfiguration: {},
+				imageTitle: nextImage.title || '',
+			},
+			[languageId]: config.adaptiveMediaEnabled
+				? nextImage
+				: nextImage.url,
+		};
+
+		delete nextEditableValue.classNameId;
+		delete nextEditableValue.classPK;
+		delete nextEditableValue.collectionFieldId;
+		delete nextEditableValue.fieldId;
+		delete nextEditableValue.mappedField;
+
+		dispatch(
+			updateEditableValuesThunk({
+				editableValues: setIn(
+					editableValues,
+					[processorKey, editableId],
+					nextEditableValue
+				),
+				fragmentEntryLinkId,
+				segmentsExperienceId,
+			})
+		);
+	};
+
+	const handleImageDescriptionChanged = (nextImageDescription) => {
+		dispatch(
+			updateEditableValuesThunk({
+				editableValues: updateIn(
+					editableValues,
+					[processorKey, editableId, 'config', 'alt'],
+					(alt) => {
+
+						// If alt is a string (old style), we need to
+						// migrate it to an object to allow translations.
+
+						if (typeof alt === 'string') {
+							return {
+								[config.defaultLanguageId]: alt,
+								[languageId]: nextImageDescription,
+							};
+						}
+
+						return {
+							...alt,
+							[languageId]: nextImageDescription,
+						};
+					},
+					{}
+				),
+				fragmentEntryLinkId,
+				segmentsExperienceId,
+			})
+		);
+	};
+
+	const handleImageSizeChanged = (imageSizeId) => {
+		dispatch(
+			updateEditableValuesThunk({
+				editableValues: setIn(
+					editableValues,
+					[
+						processorKey,
+						editableId,
+						'config',
+						'imageConfiguration',
+						selectedViewportSize,
+					],
+					imageSizeId
+				),
+				fragmentEntryLinkId,
+				segmentsExperienceId,
+			})
+		);
+	};
+
+	return (
+		<>
+			{selectedViewportSize === VIEWPORT_SIZES.desktop && (
+				<ImageSelector
+					imageTitle={imageTitle}
+					label={Liferay.Language.get('image')}
+					onClearButtonPressed={() => {
+						handleImageChanged({
+							fileEntryId: '',
+							title: '',
+							url: '',
+						});
+					}}
+					onImageSelected={handleImageChanged}
+				/>
+			)}
+
+			{config.adaptiveMediaEnabled && editableContent?.fileEntryId && (
+				<ImageSelectorSize
+					editableElement={editableElement}
+					fileEntryId={editableContent.fileEntryId}
+					imageSizeId={imageSizeId}
+					onImageSizeIdChanged={handleImageSizeChanged}
+				/>
+			)}
+
+			{selectedViewportSize === VIEWPORT_SIZES.desktop &&
+				type === EDITABLE_TYPES.image && (
+					<ImageSelectorDescription
+						imageDescription={imageDescription}
+						onImageDescriptionChanged={
+							handleImageDescriptionChanged
+						}
+					/>
+				)}
+		</>
+	);
+}
+
+DirectImagePanel.propTypes = {
 	item: getEditableItemPropTypes().isRequired,
 };
